Moved header to not have the container style
